+++ src/Service/UserResolver.php
@@ -10,17 +10,24 @@ use Heptacom\AdminOpenAuth\Contract\UserEmailInterface;
 use Heptacom\AdminOpenAuth\Contract\UserKeyInterface;
 use Heptacom\AdminOpenAuth\Contract\UserResolverInterface;
 use Heptacom\AdminOpenAuth\Contract\UserTokenInterface;
+use Heptacom\AdminOpenAuth\OpenAuth\Struct\UserStructExtension;
 use Heptacom\OpenAuth\Struct\UserStruct;
 use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityRepositoryInterface;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsAnyFilter;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
+use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\PrefixFilter;
 use Shopware\Core\Framework\Util\Random;
+use Shopware\Core\System\Language\LanguageEntity;
 use Shopware\Core\System\User\Service\UserProvisioner;
+use Shopware\Core\System\User\UserEntity;
 
 class UserResolver implements UserResolverInterface
 {
     private EntityRepositoryInterface $userRepository;
+
+    private EntityRepositoryInterface $languageRepository;
 
     private UserProvisioner $userProvisioner;
 
@@ -36,6 +43,7 @@ class UserResolver implements UserResolverInterface
 
     public function __construct(
         EntityRepositoryInterface $userRepository,
+        EntityRepositoryInterface $languageRepository,
         UserProvisioner $userProvisioner,
         LoginInterface $login,
         UserEmailInterface $userEmail,
@@ -44,6 +52,7 @@ class UserResolver implements UserResolverInterface
         ClientFeatureCheckerInterface $clientFeatureChecker
     ) {
         $this->userRepository = $userRepository;
+        $this->languageRepository = $languageRepository;
         $this->userProvisioner = $userProvisioner;
         $this->login = $login;
         $this->userEmail = $userEmail;
@@ -63,16 +72,6 @@ class UserResolver implements UserResolverInterface
             $this->userProvisioner->provision($user->getPrimaryEmail(), $password, ['email' => $user->getPrimaryEmail()]);
 
             $userId = $this->findUserId($user, $clientId, $context);
-
-            $roleIds = $user->getPassthrough()['swAclRoleIds'] ?? null;
-            if ($roleIds) {
-                $this->userRepository->update([
-                    [
-                        'id' => $userId,
-                        'aclRoles' => \array_map(static fn (string $roleId) => ['id' => $roleId], $roleIds),
-                    ],
-                ], $context);
-            }
         }
 
         $this->postUpdates($user, $userId, $state, $isNew, $clientId, $context);
@@ -102,11 +101,28 @@ class UserResolver implements UserResolverInterface
 
         $this->login->setCredentials($state, $userId, $context);
 
-        if ($isNew && !$this->clientFeatureChecker->canUsersBecomeAdmin($clientId, $context)) {
-            $this->userRepository->update([[
-                'id' => $userId,
-                'admin' => false,
-            ]], $context);
+        $userChangeSet = $this->getUserInfoChangeSet($userId, $user, $isNew, $clientId, $context);
+
+        if ($isNew) {
+            /** @var UserStructExtension|null $userExtension */
+            $userExtension = $user->getPassthrough()[UserStructExtension::class] ?? null;
+
+            if (!$userExtension) {
+                return;
+            }
+
+            $userChangeSet['admin'] = $userExtension->isAdmin();
+
+            if (!$userExtension->isAdmin()) {
+                $userChangeSet['aclRoles'] = \array_map(static fn (string $roleId) => ['id' => $roleId], $userExtension->getAclRoleIds());
+            }
+        }
+
+        if (\count($userChangeSet) > 1) {
+            // check with database if update is required
+            if ($isNew || $this->isUserChanged($userChangeSet, $context)) {
+                $this->userRepository->update([$userChangeSet], $context);
+            }
         }
     }
 
@@ -127,5 +143,84 @@ class UserResolver implements UserResolverInterface
         $criteria->addFilter(new EqualsAnyFilter('email', $emails));
 
         return $this->userRepository->searchIds($criteria, $context)->firstId();
+    }
+
+    protected function findLocaleId(string $localeCode, Context $context): ?string
+    {
+        if (empty(trim($localeCode))) {
+            return null;
+        }
+
+        $filters = [
+            new EqualsFilter('locale.code', $localeCode),
+            new PrefixFilter('locale.code', $localeCode),
+        ];
+
+        foreach ($filters as $filter) {
+            $criteria = new Criteria();
+            $criteria->addFilter($filter);
+            $criteria->setLimit(1);
+
+            $language = $this->languageRepository->search($criteria, $context)->first();
+
+            if ($language instanceof LanguageEntity) {
+                return $language->getLocaleId();
+            }
+        }
+
+        return null;
+    }
+
+    protected function getUserInfoChangeSet(string $userId, UserStruct $user, bool $isNew, string $clientId, Context $context): array
+    {
+        $userChangeSet = [
+            'id' => $userId,
+        ];
+
+        if (!$isNew && !$this->clientFeatureChecker->canKeepUserUpdated($clientId, $context)) {
+            return $userChangeSet;
+        }
+
+        $userChangeSet['firstName'] = '';
+        $userChangeSet['lastName'] = $user->getDisplayName();
+
+        if (!empty($user->getFirstName()) && !empty($user->getLastName())) {
+            $userChangeSet['firstName'] = $user->getFirstName();
+            $userChangeSet['lastName'] = $user->getLastName();
+        }
+
+        $userChangeSet['email'] = $user->getPrimaryEmail();
+        $userChangeSet['timeZone'] = $user->getTimezone();
+        $userChangeSet['localeId'] = $this->findLocaleId($user->getLocale() ?? '', $context);
+
+        return \array_filter($userChangeSet, static fn ($value) => $value !== null);
+    }
+
+    protected function isUserChanged(array $userChangeSet, Context $context): bool
+    {
+        $userId = $userChangeSet['id'] ?? null;
+
+        if (!$userId) {
+            throw new \InvalidArgumentException('Changeset must include the id.');
+        }
+
+        /** @var UserEntity|null $user */
+        $user = $this->userRepository->search(new Criteria([$userChangeSet['id']]), $context)->first();
+
+        if (!$user) {
+            return true;
+        }
+
+        foreach ($userChangeSet as $key => $newValue) {
+            try {
+                if ($user->get($key) !== $newValue) {
+                    return true;
+                }
+            } catch (\InvalidArgumentException $e) {
+                return true;
+            }
+        }
+
+        return false;
     }
 }