+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Unreleased
 
 * Added client configuration to assign default roles to new non admin users
+* Added client configuration to always update the user on login with data from the IDP
 * Changed `views/administration/heptacom-admin-open-auth/page/confirm.html.twig` to immediately close the window after local storage item was set
+* Changed `Heptacom\AdminOpenAuth\Service\UserResolver` and `Heptacom\AdminOpenAuth\Component\Provider\OpenIdConnectClient::getUser` to add more data to the user (e.g. locale and timezone)
 
 # 4.1.0